moved some code around so that each element can have its own `remove` method which should properly make that element remove itself from the ui tree
